+++ v2/pkg/engine/datasource/graphql_datasource/graphql_ws_handler.go
@@ -20,7 +20,7 @@ import (
 // it is responsible for managing all subscriptions using the underlying WebSocket connection
 // if all Subscriptions are complete or cancelled/unsubscribed the handler will terminate
 type gqlWSConnectionHandler struct {
-	// The underlying net.Conn. Only used for epoll. Should not be used to shutdown the connection.
+	// The underlying net.Conn. Only used for netPoll. Should not be used to shutdown the connection.
 	conn                          net.Conn
 	requestContext, engineContext context.Context
 	log                           abstractlogger.Logger